Marshal nulls as empty text & document data structures.
Documented the data fields in the marshal structure to avoid value
ambiguity.

NULL values for eg. identicon shouldn't serialize to flat (mpsites) file
as "(null)" but rather as empty "".

Parsing header values from the flat file that are empty yielded NULL
instead of the empty string since mpw_strtok was needlessly aggressive
on that front.